% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_targets.R
\name{extract_targets}
\alias{extract_targets}
\title{Extract Targets}
\usage{
extract_targets(features, targets)
}
\arguments{
\item{features}{a list which has been produced using the \link{integrate_peaks} function}

\item{targets}{a \code{data.frame} of MRM-MS targets to extract the the xcms processed object. \code{targets} must contain the following columns;
\itemize{
    \item{\strong{name}} the generic name of the target
    \item{\strong{mz}} the m/z value of the target (as it appears in the raw file)
    \item{\strong{rt}} the approximate retention time of the target
}}
}
\value{
a \code{data.frame} of feature values for only those which corresponds to an identified target
}
\description{
Extract Targets
}
