% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate_peaks.R
\name{integrate_peaks}
\alias{integrate_peaks}
\title{Detect and integrate MRM-MS features}
\usage{
integrate_peaks(files, phenoData, pol)
}
\arguments{
\item{files}{a character vector of \code{.mzML} files which have been convereted from MRM-MS \code{.mzML} to LC-MS style \code{.mzML} files using \code{MRMConverteR}}

\item{phenoData}{a \code{data.frame}. \code{phenoData} must contain the following columns;
\itemize{
    \item{\strong{fileName}} file name
    \item{\strong{name}} sample name
}

Providing these two columns are present; \code{phenoData} can contain as many other columns as the user wishes.}

\item{pol}{a charatcer string of either \code{1} for positive mode or \code{-1} for negative mode. If input data only contains one polarity mode, then \code{polarity} can be \code{NULL}}
}
\value{
a list of two elements
\itemize{
    \item{values} values for the integrated peak areas of all features detected
    \item{definitions} a \code{data.frame} of feature defintions. See \code{\link[xcms]{XCMSnExp-class}} for more details
}
}
\description{
Use the \code{matchedFilter} algorithim \code{xcms} for feature detection in MRM-MS which have been converted to LC-MS style \code{.mzML} files.
}
